--- a/src/huffman.ts
+++ b/src/huffman.ts
@@ -1,21 +1,26 @@
-export function generateHuffmanTable(codelenValues: Map<number, number[]>): {[key: number]: {[key: number]: number}} {
-  const codelens = codelenValues.keys();
-  let iteratorResult = codelens.next();
+export interface ICodelenValues {
+  [key: number]: number[];
+}
+export interface IHuffmanTable {
+  [key: number]: {[key: number]: number};
+}
+
+export function generateHuffmanTable(codelenValues: ICodelenValues): IHuffmanTable {
+  const codelens = Object.keys(codelenValues);
   let codelen = 0;
   let codelenMax = 0;
   let codelenMin = Number.MAX_SAFE_INTEGER;
-  while (!iteratorResult.done) {
-    codelen = iteratorResult.value;
+  codelens.forEach((key) => {
+    codelen = Number(key);
     if (codelenMax < codelen) { codelenMax = codelen; }
     if (codelenMin > codelen) { codelenMin = codelen; }
-    iteratorResult = codelens.next();
-  }
+  });
 
   let code = 0;
   let values: number[];
-  const bitlenTables: {[key: number]: {[key: number]: number}} = {};
+  const bitlenTables: IHuffmanTable = {};
   for (let bitlen = codelenMin; bitlen <= codelenMax; bitlen++) {
-    values = codelenValues.get(bitlen) as number[];
+    values = codelenValues[bitlen];
     if (values === undefined) { values = []; }
     values.sort((a, b) => {
       if ( a < b ) { return -1; }
@@ -33,16 +38,16 @@ export function generateHuffmanTable(codelenValues: Map<number, number[]>): {[ke
   return bitlenTables;
 }
 
-export function makeFixedHuffmanCodelenValues(): Map<number, number[]> {
-  const codelenValues = new Map();
-  codelenValues.set(7, new Array());
-  codelenValues.set(8, new Array());
-  codelenValues.set(9, new Array());
+export function makeFixedHuffmanCodelenValues(): ICodelenValues {
+  const codelenValues: ICodelenValues = {};
+  codelenValues[7] = [];
+  codelenValues[8] = [];
+  codelenValues[9] = [];
   for (let i = 0; i <= 287; i++) {
-    (i <= 143) ? codelenValues.get(8).push(i) :
-    (i <= 255) ? codelenValues.get(9).push(i) :
-    (i <= 279) ? codelenValues.get(7).push(i) :
-    codelenValues.get(8).push(i);
+    (i <= 143) ? codelenValues[8].push(i) :
+    (i <= 255) ? codelenValues[9].push(i) :
+    (i <= 279) ? codelenValues[7].push(i) :
+    codelenValues[8].push(i);
   }
   return codelenValues;
 }