Improve inflate speed.
Speed is measured by binary data test.

### Change the inflate table from Map to Object https://github.com/zprodev/zlib.es/pull/1/commits/75deef6e6d4f173fd86a0889c37a2c217f6d62b0

|No.|before|after|
|:---:|---:|---:|
|1|147ms|130ms|
|2|159ms|127ms|
|3|152ms|121ms|
|4|156ms|120ms|
|5|155ms|125ms|

### Change the inflate CodelenValues from Map to Object https://github.com/zprodev/zlib.es/pull/1/commits/173fe0d6b8e8d9e4c1c67078b2b4130da1aa0b34

|No.|after|
|:---:|---:|
|1|103ms|
|2|103ms|
|3|102ms|
|4|109ms|
|5|106ms|

### Refactor of iterative processing https://github.com/zprodev/zlib.es/pull/1/commits/6fe4e090e0c3ffcb525bde9e38a38ae67f5d647d

|No.|after|
|:---:|---:|
|1|99ms|
|2|101ms|
|3|109ms|
|4|107ms|
|5|103ms|